requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
pillow>=10.0.0
pytesseract>=0.3.10
plotly>=5.17.0
google-generativeai>=0.3.0
language-tool-python>=2.7.1
streamlit-lottie>=0.0.5
streamlit-player>=0.1.5
matplotlib>=3.7.0
wordcloud>=1.9.0
PyMuPDF>=1.23.0
pdfplumber>=0.10.0
reportlab>=4.0.0
requests>=2.31.0
sympy>=1.12